% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eesm_a.r
\docType{data}
\name{eesm_a}
\alias{eesm_a}
\title{Employment-related data for states and metro areas, annual.}
\format{
A data frame with approximately 500k rows; and 14 variables:
\describe{
  \item{year}{calendar year}
  \item{season}{S for seasonally adjusted, U for unadjusted}
  \item{dtype}{data type}
  \item{dtypef}{data type factor}
  \item{ssector}{super sector}
  \item{ssectorf}{super sector factor}
  \item{ind}{industry code}
  \item{indf}{industry code factor}
  \item{stabbr}{state abbreviation}
  \item{stcode}{state fips code}
  \item{stname}{state name}
  \item{area}{area code}
  \item{areaf}{area code factor}
  \item{value}{annual average value}
}
}
\source{
\url{https://download.bls.gov/pub/time.series/compressed/tape.format/}
}
\usage{
eesm_a
}
\description{
A dataset containing BLS state and metropolitan area data, annual.
}
\keyword{datasets}
