% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eesm_m.r
\docType{data}
\name{eesm_m}
\alias{eesm_m}
\title{Employment-related data for states and metro areas, monthly.}
\format{
A data frame with approximately 8 million rows; and 14 variables:
\describe{
  \item{date}{first day of month}
  \item{season}{S for seasonally adjusted, U for unadjusted}
  \item{dtype}{data type}
  \item{dtypef}{data type factor}
  \item{ssector}{super sector}
  \item{ssectorf}{super sector factor}
  \item{ind}{industry code}
  \item{indf}{industry code factor}
  \item{stabbr}{state abbreviation}
  \item{stcode}{state fips code}
  \item{stname}{state name}
  \item{area}{area code}
  \item{areaf}{area code factor}
  \item{value}{monthly value}
}
}
\source{
\url{https://download.bls.gov/pub/time.series/compressed/tape.format/}
}
\usage{
eesm_m
}
\description{
A dataset containing BLS state and metropolitan area data, monthly.
}
\keyword{datasets}
